Convert lint workflow to use pre-commit
I'm not sure if using flake8 is required for lsst_distrib projects, but
adding ruff as well as also checking against flake8 seems odd.

The cleanest solution is to having linting in GitHub Actions match the
linting behaviour that's configured by the repository and for local
development, and that's by running the pre-commit actions.