Add integration tests for packagetoctree
It's often easier to and more effective to test Sphinx extensions by
building a test site and then analyzing the HTML output to check that
the generated site is as expected. Sphinx does this itself with a Pytest
fixture that loads up different site roots in the tests/roots/
directory. Test functions get a Sphinx app instance pointing at a given
site root, and can run a site build.

Here we're both setting up the fixture and also performing integration
tests for the packagetoctree extension. We're using beautifulsoup4 to
parse and inspect the HTML output.